requirements.txt
social-auth-core==4.1.0
python-telegram-bot
mlbgame
MLB-StatsAPI